+++ packages/meditrak-server/package.json
@@ -31,6 +31,7 @@
     "http": "^0.0.0",
     "jsonwebtoken": "^7.4.3",
     "lodash": "^4.17.11",
+    "lodash.keyby": "^4.6.0",
     "moment": "^2.24.0",
     "moment-timezone": "^0.5.25",
     "multer": "^1.4.1",
