--- packages/meditrak-server/src/permissions/get/index.js
@@ -1,9 +1,0 @@
-/**
- * Tupaia
- * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
- */
-
-export {
-  hasSurveyResponsePermissions,
-  assertSurveyResponsePermissions,
-} from './hasSurveyResponsePermissions';
