+++ packages/meditrak-server/src/routes/GETSurveyResponses/index.js
@@ -1,0 +1,6 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ */
+
+export { GETSurveyResponses } from './GETSurveyResponses';
